night layout 추가, 밤 투표 결과화면 수정
